wic: Changed calls for exec_native_cmd to exec_cmd.
Since we are not making use of the official yocto oe core, we do not have native built-in commands,
within isar. Instead me have to make use of the host system commands.

At this point it may be required to run wic as super user.
sudo -s -E PATH=$PATH -- wic

Signed-off-by: Benedikt Niedermayr <Benedikt.Niedermayr@mixed-mode.de>