Introducing better debian cross-build environments. Heavy code refactoring.
Cross compilation is now done by following the debian way of running cross builds
within a cross compiling chrooted root filesystem.
Isar now offers a (hopefully) easier way when changing to a proper and complete
cross environment.

The proplem for installing inter-package build dependencies is now solved when doing
cross compiling. (When running native qemu compilations is was never a big problem.)

So the SYSROOT problem is now closer to a proper solution.
Buildependencies for cross builds where installed on the CROSS_BUILDCHROOT and
buildependencies for qemu build where installed on the BUILDCHROOT.

Closes siemens-ct/siemens-ct-ree-sdi#60
Closes siemens-ct/siemens-ct-ree-sdi#58

Signed-off-by: Benedikt Niedermayr <Benedikt.Niedermayr@mixed-mode.de>