image generation: Improved image generation.
This commit is an enhancement of creating images and describing them.
For description now a image_layout.json file has to be written. This file
describes all important attributes for each partition (label, size, mountpoint etc.) and where
in the image those partitions are located (offset or order within the image).

Signed-off-by: Benedikt Niedermayr <Benedikt.Niedermayr@mixed-mode.de>